create pool module from hash_pool.rs
